Change object types and weatherDao method to return a flowable.

+++ app/src/main/java/com/raudonikiss/weatherforecast/network/response/Clouds.kt
@@ -1,5 +1,5 @@
 package com.raudonikiss.weatherforecast.network.response
 
 data class Clouds(
-    val all: Int
+    val all: Double
 )
